{"schema":"decodex/commit/1","summary":"Split Decodex app panel support","authority":"XY-1115"}
