+++ apps/decodex-app/Sources/DecodexApp/PanelSupport.swift
@@ -1,0 +1,215 @@
+import SwiftUI
+
+enum PanelPalette {
+	static func primaryText(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.95, green: 0.96, blue: 0.98).opacity(0.97)
+			: Color(red: 0.12, green: 0.14, blue: 0.18).opacity(0.94)
+	}
+
+	static func secondaryText(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.73, green: 0.76, blue: 0.82).opacity(0.84)
+			: Color(red: 0.34, green: 0.38, blue: 0.45).opacity(0.8)
+	}
+
+	static func separator(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color.white.opacity(0.065)
+			: Color(red: 0.32, green: 0.38, blue: 0.46).opacity(0.14)
+	}
+
+	static func actionBlue(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.86, green: 0.89, blue: 0.94)
+			: Color(red: 0.18, green: 0.29, blue: 0.4)
+	}
+
+	static func codexAccent(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.82, green: 0.87, blue: 0.94)
+			: Color(red: 0.2, green: 0.36, blue: 0.52)
+	}
+
+	static func routeAccent(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.68, green: 0.8, blue: 0.96)
+			: Color(red: 0.13, green: 0.32, blue: 0.56)
+	}
+
+	static func landingAccent(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.74, green: 0.82, blue: 0.9)
+			: Color(red: 0.19, green: 0.34, blue: 0.46)
+	}
+
+	static func capacityAccent(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.72, green: 0.8, blue: 0.88)
+			: Color(red: 0.18, green: 0.34, blue: 0.48)
+	}
+
+	static func warning(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.95, green: 0.68, blue: 0.38)
+			: Color(red: 0.62, green: 0.4, blue: 0.12)
+	}
+
+	static func usageCyan(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.35, green: 0.78, blue: 0.86)
+			: Color(red: 0.1, green: 0.53, blue: 0.62)
+	}
+
+	static func fastModeAccent(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.98, green: 0.84, blue: 0.48)
+			: Color(red: 0.42, green: 0.31, blue: 0.09)
+	}
+
+	static func destructive(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color(red: 0.98, green: 0.4, blue: 0.45)
+			: Color(red: 0.68, green: 0.1, blue: 0.16)
+	}
+
+	static func progressTrack(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color.white.opacity(0.09)
+			: Color(red: 0.15, green: 0.23, blue: 0.3).opacity(0.1)
+	}
+
+	static func progressEdge(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color.white.opacity(0.12)
+			: Color.white.opacity(0.22)
+	}
+
+	static func glassStroke(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color.white.opacity(0.14)
+			: Color(red: 0.34, green: 0.42, blue: 0.52).opacity(0.24)
+	}
+
+	static func glassInnerShadow(_ colorScheme: ColorScheme) -> Color {
+		colorScheme == .dark
+			? Color.black.opacity(0.18)
+			: Color.black.opacity(0.055)
+	}
+}
+
+enum PanelMotion {
+	static let hover = Animation.interactiveSpring(response: 0.22, dampingFraction: 0.86, blendDuration: 0.04)
+	static let press = Animation.interactiveSpring(response: 0.16, dampingFraction: 0.78, blendDuration: 0.02)
+	static let state = Animation.interactiveSpring(response: 0.24, dampingFraction: 0.88, blendDuration: 0.05)
+}
+
+private extension View {
+	func panelInteractiveSurface(
+		isPressed: Bool = false,
+		isDisabled: Bool = false,
+		hoverLift: CGFloat = 0.7,
+		hoverScale: CGFloat = 1.006,
+		pressedScale: CGFloat = 0.985,
+		hoverShadowRadius: CGFloat = 3
+	) -> some View {
+		modifier(
+			PanelInteractiveSurfaceModifier(
+				isPressed: isPressed,
+				isDisabled: isDisabled,
+				hoverLift: hoverLift,
+				hoverScale: hoverScale,
+				pressedScale: pressedScale,
+				hoverShadowRadius: hoverShadowRadius
+			)
+		)
+	}
+}
+
+struct PanelInteractiveButtonStyle: ButtonStyle {
+	let isDisabled: Bool
+	let hoverLift: CGFloat
+	let hoverScale: CGFloat
+	let pressedScale: CGFloat
+	let hoverShadowRadius: CGFloat
+
+	init(
+		isDisabled: Bool = false,
+		hoverLift: CGFloat = 0.7,
+		hoverScale: CGFloat = 1.006,
+		pressedScale: CGFloat = 0.985,
+		hoverShadowRadius: CGFloat = 3
+	) {
+		self.isDisabled = isDisabled
+		self.hoverLift = hoverLift
+		self.hoverScale = hoverScale
+		self.pressedScale = pressedScale
+		self.hoverShadowRadius = hoverShadowRadius
+	}
+
+	func makeBody(configuration: Configuration) -> some View {
+		configuration.label
+			.panelInteractiveSurface(
+				isPressed: configuration.isPressed,
+				isDisabled: isDisabled,
+				hoverLift: hoverLift,
+				hoverScale: hoverScale,
+				pressedScale: pressedScale,
+				hoverShadowRadius: hoverShadowRadius
+			)
+	}
+}
+
+private struct PanelInteractiveSurfaceModifier: ViewModifier {
+	@Environment(\.colorScheme) private var colorScheme
+	@State private var isHovered = false
+	let isPressed: Bool
+	let isDisabled: Bool
+	let hoverLift: CGFloat
+	let hoverScale: CGFloat
+	let pressedScale: CGFloat
+	let hoverShadowRadius: CGFloat
+
+	func body(content: Content) -> some View {
+		let responds = isDisabled == false
+		let hoverActive = responds && isHovered && isPressed == false
+		let pressActive = responds && isPressed
+
+		content
+			.scaleEffect(pressActive ? pressedScale : (hoverActive ? hoverScale : 1))
+			.offset(y: hoverActive ? -hoverLift : 0)
+			.brightness(hoverActive ? hoverBrightness : (pressActive ? pressBrightness : 0))
+			.shadow(
+				color: hoverShadowColor.opacity(hoverActive ? 1 : 0),
+				radius: hoverActive ? hoverShadowRadius : 0,
+				x: 0,
+				y: hoverActive ? 1.8 : 0
+			)
+			.onHover { hovering in
+				guard responds else {
+					return
+				}
+
+				withAnimation(PanelMotion.hover) {
+					isHovered = hovering
+				}
+			}
+			.animation(PanelMotion.press, value: isPressed)
+			.animation(PanelMotion.hover, value: isHovered)
+			.animation(PanelMotion.state, value: isDisabled)
+	}
+
+	private var hoverBrightness: Double {
+		colorScheme == .dark ? 0.022 : 0.016
+	}
+
+	private var pressBrightness: Double {
+		colorScheme == .dark ? 0.006 : -0.004
+	}
+
+	private var hoverShadowColor: Color {
+		colorScheme == .dark
+			? Color.black.opacity(0.18)
+			: Color.black.opacity(0.09)
+	}
+}
